backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6

# 环境变量
python-dotenv==1.0.0

# AI 和机器学习 (CPU 版本)
# PyTorch CPU only - 适合无 GPU 环境
torch==2.2.0+cpu; sys_platform == 'linux'
--extra-index-url https://download.pytorch.org/whl/cpu

# 计算机视觉
numpy==1.26.3
opencv-python-headless==4.9.0.80
mediapipe==0.10.11

# YOLO 模型
ultralytics==8.1.10
onnxruntime==1.17.0

# 数据处理
pandas==2.2.0
Pillow==10.2.0

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
